ユーザ新規追加処理実装

+++ src/main/java/jp/co/mgsystems/yuricollection/gootscatalog/beans/Genre.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class Genre {
     // ジャンルID
     @NotNull
-    private Integer genreId;
+    private Long genreId;
     // ジャンル名
     private String genreName;
 }
